feat(log): o "Registros de Log" do legado — o Apollo grava e mostra a LOG (mig 313)
A LOG (2,49 mi, viva) é o histórico que 21 telas mostram pelo "Registro de log": quem inseriu ou
alterou, quando, e o valor anterior/atual de cada campo. O form-base grava uma linha a cada gravação de
cadastro (uCadMaster.pas:485) e o controle de permissões registra cada acesso liberado ou removido — o
único registro de quem mudou a permissão de quem (a audit_permissoes só guarda programa e máquina; o
dossiê de permissões a dava como cobertura, por engano).

- gravação (shared/log): o diff do form-base, comparando valor e não texto (5.50 do banco = 5,5 do
  formulário), no formato da produção (maiúsculas sem acento); IDLOG pela seq_log.
- os dois motores gravam quando o cadastro declara `log` — 18 cadastros com o título que a produção grava.
- permissões: os textos de GetMsgAcaoLog (uma permissão, todas as da tela, acesso total/geral,
  clonagem — com o defeito do legado que a produção grava).
- visualizador (uRegistrosLog): chave + valor + período (30 dias) + ação, gate da tela; modal no menu
  Outros de 12 cadastros (contas a pagar/receber e NF já mostram a história carregada) e o botão do
  controle de permissões.
- carga: o acerto genérico de sequências fazia setval(numeric) — a coluna do Oracle é numeric(10,0);
  teria parado a carga. ::bigint.

api tsc 0 · test 216 · smoke 1454/0 (+3) · web tsc 0 · test 37 · build.

+++ apps/api/scripts/carregar-cutover.ts
@@ -205,7 +205,8 @@ async function main() {
           WHERE a.attrelid = $1::regclass AND a.attnum > 0 AND NOT a.attisdropped
             AND pg_get_serial_sequence($1, a.attname) IS NOT NULL`, [tabela])).rows as Array<{ col: string; seq: string }>;
       for (const c of cols) {
-        await pool.query(`SELECT setval($1, coalesce((SELECT max(${c.col}) FROM ${tabela}), 0) + 1, false)`, [c.seq]);
+        // ::bigint — a coluna pode ser numeric(10,0) (tipo do Oracle, mig 311: log.idlog) e setval não aceita numeric
+        await pool.query(`SELECT setval($1, (coalesce((SELECT max(${c.col}) FROM ${tabela}), 0) + 1)::bigint, false)`, [c.seq]);
         seqs++;
       }
     }
